VOCA-16: add tests for dictionary dao methods

+++ backend/src/test/java/ru/jpol/vocabot/Constants.java
@@ -1,7 +1,7 @@
 package ru.jpol.vocabot;
 
 public interface Constants {
-    long defaultListSize = 5;
+    int defaultListSize = 5;
 
     Long defaultChatID = 0L;
